Add 'sv' property to 'sourceSet'
Currently, 'sv' is a 'SourceDirectorySet', because this class provides
most of the methods we need (setting the directory, filtering, etc.).
This class also has a lot of Java related methods as well, so we should
eventually write our own specialized class.

The test for the source set also went into the direction of checking
behavior instead of structure.