Add 'c' source directory set property to source sets
Some libraries make use of DPI or VPI code, which is generally written
in C. Xcelium, for example, accepts also C code.

In the long run, C code could be handled by the "native" plugin, which
already comes with Gradle. This would be a bit more involved, though,
because it would require configuration of the GCC compiler, the
include paths from the EDA tool (for 'vpi.h', etc.), etc.