Implement execution of 'runSVUnit', gotten as source dependency
Define a configuration to store the path to the downloaded SVUnit repo.
Projects using the 'svunit' plugin must define a dependency on SVUnit.
This is how the 'java' plugin also works - there is no default
dependency of JUnit (probably).

Projects using the 'svunit' plugin must define a 'sourceControl' mapping
with the path to the SVUnit repo. Since SVUnit isn't built with Gradle,
we need to inject a plugin. This plugin will simply export the root
SVUnit directory as an artifact, to be used by 'TestTask'.

Execute 'runSVUnit' in a shell, after sourcing the setup script.

This entire functionality is difficult to test, at least at the
moment, so it has minimal test coverage.