Split 'svunit-build' plugin into 'Project' and 'Settings' parts
The 'svunit-build' plugin is a 'Project' plugin that should be applied
to SVUnit. This should eventually be applied in the SVUnit repo itself,
once it has a Gradle setup.

The 'svunit-build-injector' plugin is a 'Settings' plugin. Its only task
is to inject the 'svunit-build' plugin. This will become obsolete once
we start applying 'svunit-build' in SVUnit itself.