Add `genQrunArgsFile` task
Currently, this is just a copy of the `genArgsFile` task, which
generates an args file for Xcelium. This will be updated to generate the
proper syntax for QuestaSim.

In the long term, there should also be an explicitly named task that
generates an args file for Xcelium and the `genArgsFile` task should
call either the one for Xcelium or for QuestaSim, according to a defined
selection policy.