Add example that runs in QuestaSim
It's currently using the output of the `genQrunArgsFile` directly. It
will have to use the output of a `genFullQrunArgsFile` task once this is
added.

Even though we haven't yet added support for include directories, the
code does compile in QuestaSim. This is because the parent directory of
a compiled file is always used as an include directory (as opposed to
other tools).