Fix handling of transitive dependencies
We have a very poor handling of transitive dependencies at the moment.
This is because we don't have any connection between the artifacts a
project exports (i.e. the "simple" `args.f` files) and the ones of
dependencies. Gradle hasn't been told that in order to build `args.f`
for the current project, it should build `args.f` for all dependencies
as well. The user guide states to avoid configurations that are both
consumable and resolvable, but we don't have a choice at the moment.
(Also, this is something we were doing before with the `compile`
configuration.)