Switch to using `jgitver` plugin to extract version from Git
This plugin supports `v` as a prefix for tags, as recommended by GitHub
and removes it when computing the version. The `git-version` plugin
doesn't support this, because Palantir is opposed to the prefix.

Whether using the `v` prefix or not is a good idea is unclear. In any
case we started with this scheme, so unless we have a good reason to
deviate we should stick with it.