Create compile configuration for each source set
The compile configuration of a source set is used to specify its
dependencies. This allows us to more finely model dependencies between
source sets.

It's not clear why Gradle doesn't understand that the *compile*
configuration should also depend on the output of the *prePreMain*
source set. Since it depends on the output of the *preMain* source set,
which itself depends on the output of the *prePreMain* source set,
Gradle should have been able to figure it out.