Use lazy configuration for `runSVUnit` toolchain `args`
This makes it possible to add `args` at execution time, when `testWith*`
is called from a different task. It is also the recommended way of
exposing properties.

This should be mostly backwards compatible, as the Groovy DSL adds the
necessary convenience methods to setting `Property` objects look just
like "regular" properties. (Indeed, the only change we had to make was
in the testing code, where we call `get()`. User code will most likely
not do this.)