Fix crash when specifying test compile dependency on source set
The `getGroup()` and `getName()` methods may return `null`, for example
in our case, where we don't specify a dependency on a module using
`"<group>:<name>"` coordinates.

It's a common Java idiom to call `equals()` on the string constants,
which we know are never `null`, instead of adding `null` checks. This
makes for more compact code.