Add task that zips main SystemVerilog source code into archive
This will be part of the artifact that will be published by the project. The other part, aside from sources, will be metadata necessary to compile them.

Currently, the task assumes that the SV source is in the conventional location. This will be fixed later to work in the general case.