Remove usage of non-public API when defining component
Defining custom components is possible using the public API: https://docs.gradle.org/8.6/userguide/publishing_customization.html#sec:publishing-custom-components.

The name of the variant is not configurable anymore. It's just going to be the name of the configuration we used to define the variant. This is not a big deal, as the name is not used anywhere, only its attributes matter.