Fix unzip directory not getting recreated after cleaning build directory
Why this happens isn't really clear. Gradle shows that the transforms are always executed, always when the `GenFullXrunArgsFile` task executes. It might not be a good idea to write transform outputs to a directory that is not managed by that transform (i.e. declared using `outputs.dir()`), because Gradle has no way of tracking that we did that.

We can't use the path we tried to allocate in the build directory as an argument for `outputs.dir()`, because absolute paths are only allowed to point to the `InputArtifact`. The only option we have is to chain two transforms.