Generate args file using SystemVerilog source files listed in compile spec
As a first step, we can do some "raw" XML parsing. In the long term, this should be fused with the complementary code that handles writing of compile spec XMLs.

Enabling debug will cause the `GradleRunner` to run in the same process as the `Test` task and hence run with assertions enabled.