Fix build to properly load JAXB on JDK9 and upward
JAXB is only available in the JDK until version 8. In JDK9 it is still technically available, but has to be activated using a compile/runtime switch, which is not compatible with JDK8. Starting with JDK11, the package was removed altogether and has to be procured as a 3rd party library.

The tests work without the dependency, because they run in JDK8. Trying to run a "real" Gradle build might use a newer JDK, which would cause errors due to missing classes.