Fix mysterious classloader issue in JDK11 related to JAXB
Code in the `WriteXrunArgsFile` artifact transform fails with a classloader error in JDK11:

```
         > javax.xml.bind.JAXBException: Implementation of JAXB-API has not been found on module path or classpath.
            - with linked exception:
           [java.lang.ClassNotFoundException: com.sun.xml.bind.v2.ContextFactory]
```

The `WriteCompileSpecFile` task also uses `JAXBException`, but it has no problem in finding the implementation classes. The class mentioned in the error message is present in the JAR. Why it fails is a mystery.

Version 3 of JAXB works in both JDK8 and JDK11. The namespace has changed from `javax` to `jakarta`, though. Version 4 requires JDK11, so we can't use it because we have to keep supporting JDK8.