Fix handling of SV private header dirs to support any location in project tree
We use the same interpretation as when producing args files: the `srcDirs` of the source directory set are the directories to include. We still have to take all files in each of these directories, because we don't have a way to track the private headers yet.

The tricky part here is to make this work in a "lazy" way. We can't just call `svSourceSet.getSv().getSrcDirs()` and loop over those directories, because then we would only ever get the default directory (i.e.e "src/main/sv"). We have to map each source directory to `project.files()`, which will then give us all files under that directory. Since we use providers, we are just defining the mapping, but the actual computation is done when the task gets executed.