Fix serialization/deserialization of C sources to JSON
The wonky name of the getter, with the double capital letter, confuses Jackson, causing it to think that the property is called `csourceFiles` (not the lowercase "s"). It needs the annotation both for the getter, which is used for serialization, and for the field, which is (probably) used for deserialization.