Add simplified infrastructure to configure files that should be compiled first
It's trickier to implement support for the `order` closure. This easier to implement DSL gives us the chance to start working on a real implementation of ordering.

For the moment, it's acceptable to take a shortcut and extend an internal class, because using the decorator pattern to implement `DefaultSystemVerilogSourceDirectorySet` is extremely tedious.