Fix brittle test which relied on output format
The `properties` task used to output a line with `plugin: pluginClassA, pluginClassB, ...`. With the newer Gradle version, it only outputs `plugins: Plugin Collection` (i.e. the name of the container, not the contents).

Relying on output format is brittle. It's more stable to explicitly print what we're interested in and capture that.