Fix race condition for genArgsFile tasks related to reconfiguring the build directory
The value of the `destination` property is set using the default value of the project's `buildDir` property, when the plugin is applied. If the `buildDir` property is changed afterwards, the `destination` property will not be updated. Using a lazy property fixes it, because the value is evaluated after everything is configured.

This also improves compatibility with Gradle 9, as the `buildDir` property is deprecated.